requirements.txt
appdirs==1.4.3
argcomplete==1.8.2
backports.shutil-get-terminal-size==1.0.0
base58==0.2.4
beautifulsoup4==4.5.3
boto==2.46.1
boto3==1.4.4
botocore==1.5.31
click==6.7
decorator==4.0.11
Django==1.8.17
django-appconf==1.0.2
django-cors-headers==2.0.0
django-forms-bootstrap==3.0.1
django-imagekit==3.3
django-pagination-py3==1.1.2
-e git+https://github.com/eluttine/django-planet@e592c9ac81528f6dafdb5a9ef40c711f243fcd5c#egg=django_planet
django-storages==1.5.2
django-tagging==0.3.6
djangorestframework==3.5.3
docutils==0.13.1
durationpy==0.3
enum34==1.1.6
feedparser==5.2.1
futures==3.0.5
hjson==2.0.2
ipdb==0.10.2
ipython==5.3.0
ipython-genutils==0.2.0
jmespath==0.9.2
kappa==0.6.0
lambda-packages==0.14.2
olefile==0.44
packaging==16.8
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
pilkit==2.0
Pillow==4.0.0
pinax-theme-bootstrap==3.0a4
pkg-resources==0.0.0
placebo==0.8.1
prompt-toolkit==1.0.13
psycopg2==2.7.1
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
python-slugify==1.2.1
PyYAML==3.12
requests==2.13.0
s3transfer==0.1.10
scandir==1.5
simplegeneric==0.8.1
six==1.10.0
toml==0.9.2
tqdm==4.11.2
traitlets==4.3.2
troposphere==1.9.2
Unidecode==0.4.20
w3lib==1.14.1
wcwidth==0.1.7
Werkzeug==0.12
wsgi-request-logger==0.4.6
zappa==0.40.0
